+++ CHANGELOG.md
@@ -5,8 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.3.0] - 2018-05-08
 ### Added
 - auto login pipeline for users to be able to login automatically after registering in BigCommerce store.
+- getUser pipeline providing only user id.
 
 ## [0.2.0] - 2018-04-17
 ### Added
@@ -16,6 +19,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 - login pipeline now calls an additional step at the end (from the bigcommerce cart extension) that merges the anonymous cart into the logged in cart
 
-[Unreleased]: https://github.com/shopgate/ext-bigcommerce-user/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/shopgate/ext-bigcommerce-user/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/shopgate/ext-bigcommerce-user/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/shopgate/ext-bigcommerce-user/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/shopgate/ext-bigcommerce-user/tree/v0.1.0